Add auto-stash to pre-commit
Unstaged files in the working directory can infuence test results when
testing a commit. This is unwanted behaviour, because we want to test
the commit only. To solve this, we can store the unstaged files in a
stash, so they are no longer in the working directory when tests are
run. That is exactly what this commit does: it stashes the working
directory in a timestamped stash. It unstashes after tests are run.