Add tests for git hooks for R
Tests are a great way to find bugs and to make sure your new code does
not break stuff. This commit adds tests for the pre-commit and
post-commit hooks. It runs slowly and poorly because the actual hooks
are run, so the R tests (R CMD check) are invoked. R CMD check is kind
of heavy/slow, so this takes some time.